requirements.txt
pandas==2.0.1
python-dotenv==1.0.0
Requests==2.29.0
